Time: 39 ms (5.24%) | Memory: 39.3 MB (5.63%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.cpp
@@ -12,7 +12,7 @@ public:
         long long low=1;
         int ans=0;
         while(low<=high){
-            long long mid=low+(high-low)/2;
+            int mid=low+(high-low)/2;
             bool z=okay(mid,h,piles);
             if(z){
                 ans=mid;
